# docs/openapi-v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "APILFBI",
    "description": "API de gestión documental CRM ↔ Laserfiche 11.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5080/"
    }
  ],
  "paths": {
...
    "/api/v1/catalogos/{nombre}": {
      "get": {
        "tags": [
          "Catalogos"
        ],
        "summary": "Listas para los campos del CRM (criterio 10): estados-documento, tipos-rechazo, etapas-rechazo,\r\ntipos-identificacion, segmentaciones, tipos-archivo, tipos-cliente.",
        "parameters": [
          {
            "name": "nombre",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaApiOfIReadOnlyListOfItemCatalogoDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaApiOfIReadOnlyListOfItemCatalogoDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaApiOfIReadOnlyListOfItemCatalogoDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/clientes/natural/documentos/consulta": {
      "post": {
        "tags": [
          "Consultas"
        ],
        "summary": "Criterio 7: cliente natural. Envíe cif, o tipoIdentificacion + noIdentificacion.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsultaNatural"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsultaNatural"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ConsultaNatural"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaApiOfConsultaNaturalRespuesta"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaApiOfConsultaNaturalRespuesta"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaApiOfConsultaNaturalRespuesta"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/clientes/juridico/documentos/consulta": {
      "post": {
        "tags": [
          "Consultas"
        ],
        "summary": "Criterio 8: cliente jurídico. Envíe cif o rtn.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsultaJuridica"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsultaJuridica"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ConsultaJuridica"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaApiOfConsultaJuridicaRespuesta"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaApiOfConsultaJuridicaRespuesta"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaApiOfConsultaJuridicaRespuesta"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/expedientes/{idExpediente}/documentos": {
      "get": {
        "tags": [
          "Documentos"
        ],
        "summary": "Documentos del expediente, igual que las consultas 7 y 8 pero directo por idExpediente.",
        "parameters": [
          {
            "name": "idExpediente",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaApiOfDocumentosExpedienteRespuesta"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaApiOfDocumentosExpedienteRespuesta"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaApiOfDocumentosExpedienteRespuesta"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cargas/{correlativo}": {
      "get": {
        "tags": [
          "Documentos"
        ],
        "summary": "Estado de una carga hecha por SFTP: Importado (con idDocumento) o Rechazado (con motivo).",
        "parameters": [
          {
            "name": "correlativo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaApiOfCargaDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaApiOfCargaDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaApiOfCargaDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documentos/{idDocumento}/contenido": {
      "get": {
        "tags": [
          "Documentos"
        ],
        "summary": "Criterio 9: el documento en stream, para que el CRM lo muestre (Content-Disposition inline).\r\nidDocumento es el ID único de Laserfiche. Admite el header Range.",
        "parameters": [
          {
            "name": "idDocumento",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FileStreamResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileStreamResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileStreamResult"
                }
              }
            }
          },
          "206": {
            "description": "Partial Content"
          }
        }
      }
    },
    "/api/v1/documentos/estado": {
      "post": {
        "tags": [
          "Estados"
        ],
        "summary": "Criterio 10: cambia el estado de un lote de documentos del expediente.\r\nPor defecto es todo o nada; con permitirParcial = true aplica los válidos (código 3).\r\nCada documento trae su propio código. El estado, el tipo de rechazo, la etapa y el ejecutivo\r\nasignado se reflejan en los campos de Laserfiche en segundos (el ejecutivo dispara el workflow).\r\nRequiere X-Operation-User; acepta Idempotency-Key.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SolicitudCambioEstado"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SolicitudCambioEstado"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SolicitudCambioEstado"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaApiOfRespuestaCambioEstado"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaApiOfRespuestaCambioEstado"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaApiOfRespuestaCambioEstado"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/expedientes": {
      "post": {
        "tags": [
          "Expedientes"
        ],
        "summary": "Paso previo a la carga por SFTP: obtiene el expediente del conjunto de llaves o lo crea.\r\nDevuelve 201 si lo creó y 200 si ya existía. El idExpediente es el que va en el XML de carga.\r\nRequiere el header X-Operation-User; acepta Idempotency-Key.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SolicitudExpediente"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SolicitudExpediente"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SolicitudExpediente"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaApiOfExpedienteRespuesta"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaApiOfExpedienteRespuesta"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaApiOfExpedienteRespuesta"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaApiOfExpedienteRespuesta"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaApiOfExpedienteRespuesta"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaApiOfExpedienteRespuesta"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/expedientes/{idExpediente}": {
      "get": {
        "tags": [
          "Expedientes"
        ],
        "summary": "Expediente con sus llaves vigentes.",
        "parameters": [
          {
            "name": "idExpediente",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaApiOfExpedienteDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaApiOfExpedienteDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaApiOfExpedienteDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConsultaJuridica": {
        "required": [
          "idTipoExpediente",
          "noCaso",
          "cif",
          "rtn"
        ],
        "type": "object",
        "properties": {
          "idTipoExpediente": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "noCaso": {
            "type": [
              "null",
              "string"
            ]
          },
          "cif": {
            "type": [
              "null",
              "string"
            ]
          },
          "rtn": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ConsultaNatural": {
        "required": [
          "idTipoExpediente",
          "noCaso",
          "cif",
          "tipoIdentificacion",
          "noIdentificacion"
        ],
        "type": "object",
        "properties": {
          "idTipoExpediente": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "noCaso": {
            "type": [
              "null",
              "string"
            ]
          },
          "cif": {
            "type": [
              "null",
              "string"
            ]
          },
          "tipoIdentificacion": {
            "type": [
              "null",
              "string"
            ]
          },
          "noIdentificacion": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "EntityTagHeaderValue": {
        "type": "object",
        "properties": {
          "tag": {
            "$ref": "#/components/schemas/StringSegment"
          },
          "isWeak": {
            "type": "boolean"
          }
        }
      },
...
      "FileStreamResult": {
        "type": "object",
        "properties": {
          "fileStream": {
            "$ref": "#/components/schemas/Stream"
          },
          "contentType": {
            "type": [
              "null",
              "string"
            ]
          },
          "fileDownloadName": {
            "type": [
              "null",
              "string"
            ]
          },
          "lastModified": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "entityTag": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EntityTagHeaderValue"
              }
            ]
          },
          "enableRangeProcessing": {
            "type": "boolean"
          }
        }
      },
      "ItemCatalogoDto": {
        "required": [
          "codigo",
          "nombre"
        ],
        "type": "object",
        "properties": {
          "codigo": {
            "type": "string"
          },
          "nombre": {
            "type": "string"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "RespuestaApiOfCargaDto": {
        "required": [
          "codigo",
          "mensaje",
          "correlationId",
          "data"
        ],
        "type": "object",
        "properties": {
          "codigo": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "mensaje": {
            "type": "string"
          },
          "correlationId": {
            "type": [
              "null",
              "string"
            ]
          },
          "data": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CargaDto"
              }
            ]
          }
        },
        "description": "Sobre estándar de las respuestas exitosas."
      },
      "RespuestaApiOfConsultaJuridicaRespuesta": {
        "required": [
          "codigo",
          "mensaje",
          "correlationId",
          "data"
        ],
        "type": "object",
        "properties": {
          "codigo": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "mensaje": {
            "type": "string"
          },
          "correlationId": {
            "type": [
              "null",
              "string"
            ]
          },
          "data": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ConsultaJuridicaRespuesta"
              }
            ]
          }
        },
        "description": "Sobre estándar de las respuestas exitosas."
      },
      "RespuestaApiOfConsultaNaturalRespuesta": {
        "required": [
          "codigo",
          "mensaje",
          "correlationId",
          "data"
        ],
        "type": "object",
        "properties": {
          "codigo": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "mensaje": {
            "type": "string"
          },
          "correlationId": {
            "type": [
              "null",
              "string"
            ]
          },
          "data": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ConsultaNaturalRespuesta"
              }
            ]
          }
        },
        "description": "Sobre estándar de las respuestas exitosas."
      },
      "RespuestaApiOfDocumentosExpedienteRespuesta": {
        "required": [
          "codigo",
          "mensaje",
          "correlationId",
          "data"
        ],
        "type": "object",
        "properties": {
          "codigo": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "mensaje": {
            "type": "string"
          },
          "correlationId": {
            "type": [
              "null",
              "string"
            ]
          },
          "data": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DocumentosExpedienteRespuesta"
              }
            ]
          }
        },
        "description": "Sobre estándar de las respuestas exitosas."
      },
...
      "RespuestaApiOfExpedienteRespuesta": {
        "required": [
          "codigo",
          "mensaje",
          "correlationId",
          "data"
        ],
        "type": "object",
        "properties": {
          "codigo": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "mensaje": {
            "type": "string"
          },
          "correlationId": {
            "type": [
              "null",
              "string"
            ]
          },
          "data": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ExpedienteRespuesta"
              }
            ]
          }
        },
        "description": "Sobre estándar de las respuestas exitosas."
      },
      "RespuestaApiOfIReadOnlyListOfItemCatalogoDto": {
        "required": [
          "codigo",
          "mensaje",
          "correlationId",
          "data"
        ],
        "type": "object",
        "properties": {
          "codigo": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "mensaje": {
            "type": "string"
          },
          "correlationId": {
            "type": [
              "null",
              "string"
            ]
          },
          "data": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/ItemCatalogoDto"
            }
          }
        },
        "description": "Sobre estándar de las respuestas exitosas."
      },
...
      "RespuestaApiOfRespuestaCambioEstado": {
        "required": [
          "codigo",
          "mensaje",
          "correlationId",
          "data"
        ],
        "type": "object",
        "properties": {
          "codigo": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "mensaje": {
            "type": "string"
          },
          "correlationId": {
            "type": [
              "null",
              "string"
            ]
          },
          "data": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RespuestaCambioEstado"
              }
            ]
          }
        },
        "description": "Sobre estándar de las respuestas exitosas."
      },
...
      "SolicitudCambioEstado": {
        "required": [
          "idExpediente",
          "ejecutivoAsignado",
          "permitirParcial",
          "documentos"
        ],
        "type": "object",
        "properties": {
          "idExpediente": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "ejecutivoAsignado": {
            "type": [
              "null",
              "string"
            ]
          },
          "permitirParcial": {
            "type": "boolean"
          },
          "documentos": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/CambioEstadoItem"
            }
          }
        }
      },
      "SolicitudExpediente": {
        "required": [
          "idTipoExpediente",
          "llaves"
        ],
        "type": "object",
        "properties": {
          "idTipoExpediente": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "llaves": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/LlaveValor"
            }
          }
        }
      },
      "Stream": {
        "type": "string",
        "format": "binary"
      },
...
    },
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "description": "OAuth 2.0 client_credentials. El token dura 1 hora.",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "/api/v1/auth/token",
            "scopes": {
              "catalogos.leer": "catalogos.leer",
              "expedientes.escribir": "expedientes.escribir",
              "documentos.leer": "documentos.leer",
              "documentos.estado": "documentos.estado",
              "bitacora.leer": "bitacora.leer"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "oauth2": [
        "catalogos.leer",
        "expedientes.escribir",
        "documentos.leer",
        "documentos.estado",
        "bitacora.leer"
      ]
    }
  ],
...
}